Add major Product functionalities

+++ projectSAP/src/main/java/com/example/demo/products/ProductRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ProductRepository
-        extends JpaRepository<Products, Long> {
+        extends JpaRepository<Products, Integer> {
 }
